CARTOON/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pillow
opencv-python-headless
numpy
